Backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.20.0
fsspec==2025.10.0
gunicorn==23.0.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
keybert==0.9.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
preshed==3.0.12
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.1.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.4
srsly==2.5.2
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
